second ommit

+++ src/ro/siit/Vehicle.java
@@ -3,5 +3,5 @@ package ro.siit;
 public interface Vehicle {
     void start();
     void stop();
-    void drive(double n);
+    void drive(float n);
 }
